enable profiling if required, cleanup

+++ Dockerfile
@@ -7,13 +7,15 @@ FROM ${BASE_IMAGE}:${PHP_VERSION}-ubuntu
 ARG PHP_VERSION=7.4
 ARG TARGETPLATFORM
 
+ENV PHP_VERSION=${PHP_VERSION}
+
 USER root
 
-RUN set -x; apt-get update; apt install php${PHP_VERSION}-fpm msmtp curl -y && apt clean all; rm -rf /var/lib/apt/lists/* 
+RUN apt-get update; apt install php${PHP_VERSION}-fpm msmtp curl -y && apt clean all; rm -rf /var/lib/apt/lists/* 
 
 # Routine to install newrelic agent
 RUN \
-  if [ "${TARGETPLATFORM}" = "linux/arm64" ] ; then exit 0; fi ; export NR_AGENT_VERSION=$(curl https://download.newrelic.com/php_agent/release/ | grep "linux.tar" | sed -E 's/.*release\/(.+)\".*/\1/'); curl -so - https://download.newrelic.com/php_agent/release/${NR_AGENT_VERSION} | tar zxf - && \
+  if [[ "${TARGETPLATFORM}" = "linux/arm64" ]] || [[ "$(uname -m)" = "aarch64" ]]; then exit 0; fi ; export NR_AGENT_VERSION=$(curl https://download.newrelic.com/php_agent/release/ | grep "linux.tar" | sed -E 's/.*release\/(.+)\".*/\1/'); curl -so - https://download.newrelic.com/php_agent/release/${NR_AGENT_VERSION} | tar zxf - && \
   cd newrelic-php* && NR_INSTALL_SILENT=1 NR_INSTALL_USE_CP_NOT_LN=1 ./newrelic-install install && \
   rm -rf /tmp/nrinstall* && \
   echo 'newrelic.daemon.start_timeout = "5s"' >> /etc/php/${PHP_VERSION}/mods-available/newrelic.ini && \
@@ -29,16 +31,14 @@ RUN \
 
 # Routine to install Data Dog agent
 # https://docs.datadoghq.com/tracing/trace_collection/dd_libraries/php/?tab=containers
-# You must set DD_AGENT_HOST and DD_TRACE_AGENT_PORT to point at your DD 
+# You must set DD_AGENT_HOST and DD_TRACE_AGENT_PORT to point at your DD Agent
 # We also clean up whatever this config file layout is
 RUN \
   curl -LO https://github.com/DataDog/dd-trace-php/releases/latest/download/datadog-setup.php -o /tmp/datadog-setup.php && \
   php datadog-setup.php --php-bin=all && \
   rm -f /tmp/datadog-setup.php && \ 
   mv /etc/php/${PHP_VERSION}/cli/conf.d/98-ddtrace.ini /etc/php/${PHP_VERSION}/mods-available/ddtrace.ini && \
-  rm -f /etc/php/${PHP_VERSION}/fpm/conf.d/98-ddtrace.ini 
-  # ln -s /etc/php/${PHP_VERSION}/mods-available/ddtrace.ini /etc/php/${PHP_VERSION}/cli/conf.d/98-ddtrace.ini && \
-  # ln -s /etc/php/${PHP_VERSION}/mods-available/ddtrace.ini /etc/php/${PHP_VERSION}/fpm/conf.d/98-ddtrace.ini 
+  rm -f /etc/php/${PHP_VERSION}/fpm/conf.d/98-ddtrace.ini
 
 RUN \
   mkdir -p /run/php-fpm && \
@@ -57,6 +57,10 @@ RUN echo "post_max_size = ${UPLOAD_LIMIT}" >> /etc/php/${PHP_VERSION}/mods-avail
 RUN echo "upload_max_filesize = ${UPLOAD_LIMIT}" >> /etc/php/${PHP_VERSION}/mods-available/upload-limits.ini
 RUN echo "catch_workers_output = yes" >> /etc/php-fpm.d/www.conf
 RUN chown 33:33 /etc/php-fpm.d/www.conf
+RUN \
+  chown 33:33 /etc/php/${PHP_VERSION}/cli/conf.d && \
+  chown 33:33 /etc/php/${PHP_VERSION}/fpm/conf.d && \
+  chown 33:33 /etc/php/${PHP_VERSION}/mods-available
 RUN phpdismod opcache && phpenmod docker-opcache upload-limits opcache 
 RUN ln -s /usr/sbin/php-fpm${PHP_VERSION} /usr/sbin/php-fpm
 
@@ -66,4 +70,4 @@ RUN ln -s /usr/bin/msmtp /usr/sbin/sendmail && chmod +x /entrypoint.sh
 
 USER www-data
 WORKDIR /var/www/html
-ENTRYPOINT ["/entrypoint.sh"]
+CMD ["/entrypoint.sh"]
